Fixed a warning (#211)

+++ lib/SPIRV/libSPIRV/SPIRVInstruction.h
@@ -1091,7 +1091,7 @@ public:
       if (!Module->exist(Pairs[PairSize*I + getLiteralsCount()], &BB))
         continue;
 
-      for (int i = 0; i < getLiteralsCount(); ++i) {
+      for (size_t i = 0; i < getLiteralsCount(); ++i) {
         Literals.push_back(Pairs.at(PairSize*I + i));
       }
       Func(Literals, static_cast<SPIRVBasicBlock *>(BB));
